Add : User login function and create user function - Add user login function and token generation function - `auth.js` is the route for authentication - `route.js` is the route for collecting `/api` routes - `middleware/verifyToken` is the middleware for token verification - Tokens are generated and store in database and cookie - The middleware for verifing token haven't completed - `/api/test` is route used for testing token
